requirements.txt
ccxt==3.0.63
python-telegram-bot==20.3
aiohttp==3.8.5